Prevent attaching comment to non-printable child of shorthand `{Import,Export}Specifier`
